fix(routing): close pool on NotALeader to re-acquire leader connection (#270)
